# docs/swagger.yaml
definitions:
  debt.BodyNewDebt:
    properties:
      description:
        type: string
      dueDate:
        example: "2024-12-25"
        type: string
      name:
        type: string
      value:
        type: number
    type: object
  debt.BodyUpdateDebt:
    properties:
      description:
        type: string
      dueDate:
        example: "2024-12-25"
        type: string
      id:
        type: string
      name:
        type: string
      paymentDate:
        example: "2024-12-25"
        type: string
      value:
        type: number
    type: object
  debt.BodyUpdateDebtPaymentStatus:
    properties:
      id:
        type: string
      paid:
        type: boolean
    type: object
  income.BodyNewIncome:
    properties:
      description:
        type: string
      name:
        type: string
      receivedDate:
        example: "2024-12-25"
        type: string
      value:
        type: number
    type: object
  income.BodyUpdateIncome:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      receivedDate:
        example: "2024-12-25"
        type: string
      value:
        type: number
    type: object
info:
  contact: {}
paths:
  /balance:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Show balance
      tags:
      - balance
  /debt:
    get:
      consumes:
      - application/json
      description: Lists all debts
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Show debts
      tags:
      - debts
    post:
      consumes:
      - application/json
      description: Creates a new debt
      parameters:
      - description: Request body to add a debt
        in: body
        name: debt
        required: true
        schema:
          $ref: '#/definitions/debt.BodyNewDebt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a debt
      tags:
      - debts
    put:
      consumes:
      - application/json
      description: Updates an existing debt
      parameters:
      - description: Request body to update debt
        in: body
        name: debt
        required: true
        schema:
          $ref: '#/definitions/debt.BodyUpdateDebt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a debt
      tags:
      - debts
  /debt/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing debt
      parameters:
      - description: Debt id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a debt
      tags:
      - debts
  /debt/payment/:
    post:
      consumes:
      - application/json
      description: Sets an existing debt paid/unpaid
      parameters:
      - description: Request body to update a debt payment status
        in: body
        name: debt
        required: true
        schema:
          $ref: '#/definitions/debt.BodyUpdateDebtPaymentStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Set a debt paid/unpaid
      tags:
      - debts
  /income:
    get:
      consumes:
      - application/json
      description: Lists all incomes
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Show incomes
      tags:
      - incomes
    post:
      consumes:
      - application/json
      description: Creates a new income
      parameters:
      - description: Request body to add an income
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/income.BodyNewIncome'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create an income
      tags:
      - incomes
    put:
      consumes:
      - application/json
      description: Updates an existing income
      parameters:
      - description: Request body to update an income
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/income.BodyUpdateIncome'
      produces:
      - application/json
      responses:
        "20":
          description: ""
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update an income
      tags:
      - incomes
  /income/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing income
      parameters:
      - description: Income id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete an income
      tags:
      - incomes
swagger: "2.0"